+++ lib/src/codecs/codecs.dart
@@ -17,3 +17,4 @@ export 'language/language_codec.dart';
 export 'g_year_month/year_month_codec.dart';
 export 'g_month/gregorian_month_codec.dart';
 export 'unsigned_short/unsigned_short_codec.dart';
+export 'positive_integer/positive_integer_codec.dart';